ml-service/app/utils.py
import numpy as np
from PIL import Image

MEAN = np.array([0.485, 0.456, 0.406])
STD = np.array([0.229, 0.224, 0.225])

def preprocess_image(image: Image.Image) -> np.ndarray:
    image = image.convert("RGB")
    image = image.resize((224, 224))
    
    img_np = np.array(image).astype(np.float32) / 255.0
    img_np = (img_np - MEAN) / STD
    img_np = img_np.transpose(2, 0, 1)
    img_np = np.expand_dims(img_np, axis=0)

    return img_np.astype(np.float32)

