ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
onnxruntime==1.20.1
pillow==10.4.0
numpy==1.26.4